--- app/components/tramite/TramiteCarrera.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import Tramite from './Tramite';
-
-const TramiteCarrera = () => {
-
-    const datos = require('../../assets/tramites.json')
-
-    return(
-        <Tramite 
-            info={datos[0]}
-        />
-    );
-
-};
-
-export default TramiteCarrera;
